requirements.txt
Django==1.11.3
djangorestframework==3.6.3
Markdown==2.6.8
psycopg2==2.7.3
PyJWT==1.5.2
pytz==2017.2
